pyproject.toml
[project]
name = "docuchat"
dynamic = ["version"]
description = "DocuChat"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "argon2-cffi>=25.1.0",
    "asgiref>=3.11.1",
    "bullmq>=2.22.0",
    "fastapi-limiter>=0.2.0",
    "fastapi[standard]>=0.135.3",
    "httpx>=0.28.1",
    "nh3>=0.3.4",
    "prometheus-client>=0.25.0",
    "pybreaker>=1.4.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "pyee>=13.0.1",
    "pyjwt>=2.12.1",
    "redis>=6.4.0",
    "secure>=2.0.1",
    "structlog>=25.5.0",
    "tortoise-orm[asyncpg]>=1.1.7",
]

[dependency-groups]
dev = ["coverage>=7.13.5", "pyright>=1.1.408", "pytest>=9.0.3", "ruff>=0.15.11"]

[tool.setuptools.dynamic.version]
attr = "app.__version__"

[tool.ruff]
fix = true
exclude = ["*/migrations"]

[tool.ruff.lint]
extend-select = [
    "UP",    # pyupgrade
    "I",     # isort
    "T20",   # flake8-print
    "ANN",   # flake8-annotations
    "ASYNC", # flake8-async
    "S",     # flake8-bandit
    "ICN",   # flake8-import-convention
    "RUF",   # ruff
    "ERA",   # eradicate
    "D",     # pydocstyle

]
ignore = ["D100", "D104", "ANN401"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.isort]
case-sensitive = true
length-sort = true


[tool.ruff.lint.extend-per-file-ignores]
"tests/*" = ["S101", "D103", "S105", "S106"]

[tool.pyright]
venvPath = "."
venv = ".venv"
include = ["app", "tests"]
ignore = ["**/migrations"]
typeCheckingMode = "strict"

useLibraryCodeForTypes = true

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.coverage.run]
branch = true
source = ["app"]
omit = ["*/migrations/*"]

[tool.coverage.report]
show_missing = true
skip_covered = true
fail_under = 0

[tool.tortoise]
tortoise_orm = "app.orm.config.TORTOISE_CONFIG"
location = "./app/orm/migrations"
src_folder = "./."
